+++ packages/ai/core/registry/provider-registry.test.ts
@@ -3,6 +3,7 @@ import { MockEmbeddingModelV1 } from '../test/mock-embedding-model-v1';
 import { MockLanguageModelV1 } from '../test/mock-language-model-v1';
 import { NoSuchProviderError } from './no-such-provider-error';
 import { experimental_createProviderRegistry } from './provider-registry';
+import { MockImageModelV1 } from '../test/mock-image-model-v1';
 
 describe('languageModel', () => {
   it('should return language model from provider', () => {
@@ -125,5 +126,52 @@ describe('textEmbeddingModel', () => {
     expect(() => registry.textEmbeddingModel('model')).toThrowError(
       NoSuchModelError,
     );
+  });
+});
+
+describe('imageModel', () => {
+  it('should return image model from provider', () => {
+    const model = new MockImageModelV1();
+
+    const modelRegistry = experimental_createProviderRegistry({
+      provider: {
+        imageModel: id => {
+          expect(id).toEqual('model');
+          return model;
+        },
+        languageModel: () => null as any,
+        textEmbeddingModel: () => null as any,
+      },
+    });
+
+    expect(modelRegistry.imageModel('provider:model')).toEqual(model);
+  });
+
+  it('should throw NoSuchProviderError if provider does not exist', () => {
+    const registry = experimental_createProviderRegistry({});
+
+    expect(() => registry.imageModel('provider:model')).toThrowError(
+      NoSuchProviderError,
+    );
+  });
+
+  it('should throw NoSuchModelError if provider does not return a model', () => {
+    const registry = experimental_createProviderRegistry({
+      provider: {
+        imageModel: () => null as any,
+        languageModel: () => null as any,
+        textEmbeddingModel: () => null as any,
+      },
+    });
+
+    expect(() => registry.imageModel('provider:model')).toThrowError(
+      NoSuchModelError,
+    );
+  });
+
+  it("should throw NoSuchModelError if model id doesn't contain a colon", () => {
+    const registry = experimental_createProviderRegistry({});
+
+    expect(() => registry.imageModel('model')).toThrowError(NoSuchModelError);
   });
 });